获取用户路径时判断工具是否存在

+++ user-dirs.sh
@@ -8,8 +8,8 @@ if [ $UID != 0 -o "$SUDO_USER" == "root" ]; then
 fi
 
 # 获取用户目录
-[ -f /bin/apt ] && apt install -y xdg-user-dirs
-[ -f /bin/pacman ] && pacman -S --noconfirm --needed xdg-user-dirs
+[ ! -f bin/xdg-user-dir ] && [ -f /bin/apt ] && apt install -y xdg-user-dirs
+[ ! -f bin/xdg-user-dir ] && [ -f /bin/pacman ] && pacman -S --noconfirm --needed xdg-user-dirs
 USER_DESKTOP=$(basename $(su - $SUDO_USER -c 'xdg-user-dir DESKTOP'))
 USER_DOWNLOAD=$(basename $(su - $SUDO_USER -c 'xdg-user-dir DOWNLOAD'))
 USER_TEMPLATES=$(basename $(su - $SUDO_USER -c 'xdg-user-dir TEMPLATES'))
